[cpu][mac] change not return error even if cpu.frequency errors
Backport of https://github.com/shirou/gopsutil/pull/1192.

sysctl hw.cpufrequency does not return anything on macOS M1. This causes cpu.Info()
to return an error every time it is called on macOS M1.

A partial fix of this issue is to ignore errors, and leave the frequency field with
its default value.